Extract recovery conditions to hook
