+++ packages/ui/src/accounts/components/AccountItem/components/LocksDetails.tsx
@@ -1,14 +1,12 @@
-import React, { useMemo } from 'react'
+import React from 'react'
 
 import { LockItemWrapper } from '@/accounts/components/AccountItem/components/LockItemWrapper'
-import { useIsVoteStakeLocked } from '@/accounts/hooks/useIsVoteStakeLocked'
+import { useRecoveryConditions } from '@/accounts/hooks/useRecoveryConditions'
 import { isRecoverable } from '@/accounts/model/lockTypes'
 import { Balances } from '@/accounts/types'
 import { RowGapBlock } from '@/common/components/page/PageContent'
 import { Label, TextMedium } from '@/common/components/typography'
 import { Address } from '@/common/types'
-import { useLatestElection } from '@/council/hooks/useLatestElection'
-import { useMyCastVotes } from '@/council/hooks/useMyCastVotes'
 
 interface LocksDetailsProps {
   balance: Balances | null
@@ -16,22 +14,7 @@ interface LocksDetailsProps {
 }
 
 export const LocksDetails = ({ balance, address }: LocksDetailsProps) => {
-  const { election } = useLatestElection()
-  const { votes } = useMyCastVotes(election?.cycleId)
-
-  const isActiveCandidate = useMemo(
-    () => election?.candidates.find((candidate) => candidate.stakingAccount === address)?.status === 'ACTIVE',
-    [election, address]
-  )
-
-  const candidate = useMemo(() => {
-    return votes?.find((vote) => vote.castBy === address)?.voteFor
-  }, [votes])
-
-  const isVoteStakeLocked = !!useIsVoteStakeLocked(candidate, {
-    isLatestElection: !!candidate,
-    isElectionFinished: election?.isFinished,
-  })
+  const { isActiveCandidate, isVoteStakeLocked } = useRecoveryConditions(address)
 
   if (!balance || !balance.locks.length) {
     return <TextMedium light>No locks found.</TextMedium>
